+++ src/titiler/application/titiler/application/__init__.py
@@ -1,3 +1,3 @@
 """titiler.application"""
 
-__version__ = "2.0.0a1"
+__version__ = "2.0.0a2"