+++ src/titiler/core/titiler/core/__init__.py
@@ -1,6 +1,6 @@
 """titiler.core"""
 
-__version__ = "2.0.0a1"
+__version__ = "2.0.0a2"
 
 from . import dependencies, errors, factory, routing  # noqa
 from .factory import (  # noqa